+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `Causer` interface
 - `WithPrefix` didn't really make sense with the decorator pattern
+- `Contextor` interface
 
 
 ## [0.7.1] - 2018-04-27